requirements.txt
Flask
pandas
matplotlib
mysql-connector-python
openpyxl
pymongo